user-microservice/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.29
pydantic==2.6.4
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-dotenv==1.0.1
alembic==1.13.1
pymysql==1.1.0
cryptography